added libyaml dependency for configs

+++ include/OccupancyGrid.h
@@ -29,8 +29,8 @@ struct OccupancyGrid {
    * @param yCoordinate is the y coordinate in m
    */
   void setAsObstacle(float xCoordinate, float yCoordinate) {
-    int xIndex = int(std::floor(xCoordinate)/resolution_m);
-    int yIndex = int(std::floor(yCoordinate)/resolution_m);
+    int xIndex = static_cast<int>(std::floor(xCoordinate / resolution_m));
+    int yIndex = static_cast<int>(std::floor(yCoordinate / resolution_m));
 
     int obsIndex = get1DIndex(xIndex, yIndex);
 
@@ -46,22 +46,22 @@ struct OccupancyGrid {
    * @param obstacle_width is the size in x direction
    */
   void setAsObstacleWithDimension(float xCoordinate, float yCoordinate, float obstacle_width, float obstacle_height) {
-    int xIndex = int(std::floor(xCoordinate)/resolution_m);
-    int yIndex = int(std::floor(yCoordinate)/resolution_m);
-    int cell_width = int(std::floor(obstacle_width)/resolution_m);
-    int cell_height = int(std::floor(obstacle_height)/resolution_m);
+    int xIndex = static_cast<int>(std::floor(xCoordinate / resolution_m));
+    int yIndex = static_cast<int>(std::floor(yCoordinate / resolution_m));
+    int cell_width = static_cast<int>(std::floor(obstacle_width / resolution_m));
+    int cell_height = static_cast<int>(std::floor(obstacle_height / resolution_m));
 
-    std::vector<float> corners_x = {
-      xIndex - cell_width / 2, // Bottom-left corner
-      xIndex + cell_width / 2, // Bottom-right corner
-      xIndex + cell_width / 2, // Top-right corner
-      xIndex - cell_width / 2  // Top-left corner
+    std::vector<int> corners_x = {
+      xIndex - cell_width / 2,
+      xIndex + cell_width / 2,
+      xIndex + cell_width / 2,
+      xIndex - cell_width / 2,
     };
-    std::vector<float> corners_y = {
-      yIndex - cell_height / 2, // Bottom-left corner
-      yIndex - cell_height / 2, // Bottom-right corner
-      yIndex + cell_height / 2, // Top-right corner
-      yIndex + cell_height / 2  // Top-left corner
+    std::vector<int> corners_y = {
+      yIndex - cell_height / 2,
+      yIndex - cell_height / 2,
+      yIndex + cell_height / 2,
+      yIndex + cell_height / 2,
     };
 
     for (int i = corners_x[0];i<=corners_x[1];i++){
